Re-worked the online bug tracking tools quite a bit, made the HTML much nicer. Fixed a bunch of small bugs that players had found...

+++ mud/grrmud/server/critter.cc
@@ -2714,7 +2714,7 @@ void critter::doSuicide() {
 
 
 void critter::setMode(PcMode val) {
-   mudlog << "Setting mode to:  " << val << endl;
+   mudlog << "Setting mode to:  " << (int)val << endl;
    pc->mode = val;
 }
